requirements.txt
kafka-python
redis
pandas
tqdm
